* Added BASIC re-implementation of BailError that is compatible with react 19's suspense.
Note: It basically just makes the rethrown bail-error look like a promise, for react-suspense to accept it. This is arguably not as elegant as returning an actual promise that resolves on the given data being loaded. (I say arguably, because mobx currently fulfills the same role by rerendering when the accessed tree-node's status/data changes; but using an actual promise like react-suspense expects may have some benefits for rendering performance or what-not that are not immediately obvious)