* Fixed some minor bugs. * Fixed that observer_mgl was not reliably rerendering when the db-data/mobx-fields the comp was subscribed to changed! (The problem is that since the fallback-comp and regular-render-comp are different, the fallback comp does not itself have mobx-observing of the relevant data; this code change makes-so the fallback-comp calls the regular-render-comp's function in order to register those accesses.)
Note: While the above *technically fixes* the not-rerendering issue, the problem is that react complains about the "order of hooks" changing (ie. when a mobx-access is able to proceed, it thinks there's a "new hook" when the next hook is reached). So I'll need to find a better fix for the problem, that integrates the mobx-graphlink "tree" structures/tracking with the fake-promise "then" functions directly.